Get Authors Papers

Get Authors Papers

Usage

get_authors_papers(id_type = c("orcid", "openalex", "scopus", "mag"), id)

Arguments

id_type

the type of id you are using to locate the author: must be either orcid or openalex

id

the id for the author that matches the type selected initially

Value

dataframe with all papers from that author, includes variables: id: openalex id of the paper display_name: title of the paper publication date: date of publication in YYYY-MM-DD form relevance score: ?Not sure what this is, all NA for me atm authorships: nested dataframe of all authors in the paper (includes author.id, author_position, author.display_name, and author.orcid) concepts: nested dataframe of the concepts associated with the paper (includes openalex's id of the concept, display.name of the concept, score of the concept, level of the concept in the concept hierarchy, and wikidata of the concept) alternate_host_venues: ?Not sure what this is either, will need to check on openalex's page cited_by_count: number of other papers that have cited the paper publication_year: int cited_by_api_url: openalex api url for the papers that have cited this paper doi: doi of the paper type: type of paper (e.g. proceedings-article, journal-article, book-chapter, posted-content) is_paratext: ?Not sure what this is either, will need to check on openalex's page is_retracted: TRUE/FALSE whether paper has been retracted referenced_works: list of all works referenced in paper (with openalex urls :) ) related_works: similar works? not sure how these are found

Examples

get_authors_papers(id_type = "openalex", id = "A2941293769")
#>                                 id
#> 1 https://openalex.org/W3041426895
#> 2 https://openalex.org/W2991581987
#> 3 https://openalex.org/W2941505258
#>                                                                                                                               display_name
#> 1                                         The Social Psychology of Occupational Status Groups: Relationality in the Structure of Deference
#> 2 Assessing the Russian Internet Research Agency's impact on the political attitudes and behaviors of American Twitter users in late 2017.
#> 3                      Extracurricular Activities and Student Outcomes in Elementary and Middle School: Causal Effects or Self-selection?:
#>   publication_date relevance_score                   host_venue.id
#> 1       2020-07-08              NA https://openalex.org/V131591925
#> 2       2020-01-07              NA https://openalex.org/V125754415
#> 3       2019-04-23              NA                            <NA>
#>                                                           host_venue.display_name
#> 1                                                     Social Psychology Quarterly
#> 2 Proceedings of the National Academy of Sciences of the United States of America
#> 3                               Socius: Sociological Research for a Dynamic World
#>                              host_venue.publisher      host_venue.issn
#> 1                                            SAGE 1939-8999, 0190-2725
#> 2 Proceedings of the National Academy of Sciences 0027-8424, 1091-6490
#> 3       SAGE PublicationsSage CA: Los Angeles, CA                 NULL
#>   host_venue.issn_l host_venue.type
#> 1         0190-2725       publisher
#> 2         0027-8424       publisher
#> 3              <NA>       publisher
#>                                                                                                                                  host_venue.url
#> 1 https://sage.figshare.com/collections/The_Social_Psychology_of_Occupational_Status_Groups_Relationality_in_the_Structure_of_Deference/5054390
#> 2                                                                                                       https://doi.org/10.1073/pnas.1906420116
#> 3                                                                                                      https://doi.org/10.1177/2378023119845496
#>   host_venue.is_oa host_venue.version host_venue.license
#> 1               NA               <NA>               <NA>
#> 2             TRUE   publishedVersion        cc-by-nc-nd
#> 3             TRUE   publishedVersion           cc-by-nc
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         authorships
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          https://openalex.org/A2941293769, Emily Maloney, NA, https://openalex.org/I170897317, Duke University, https://ror.org/00py81415, US, education, first, Duke University, Durham, NC, USA
#> 2 https://openalex.org/A2557604234, https://openalex.org/A2296125463, https://openalex.org/A2941293769, https://openalex.org/A2989707772, https://openalex.org/A2254831310, https://openalex.org/A2807433028, https://openalex.org/A2079431213, https://openalex.org/A1611495263, Christopher A. Bail, Brian Guay, Emily Maloney, Aidan Combs, D. Sunshine Hillygus, Friedolin Merhout, Deen Freelon, Alexander Volfovsky, https://orcid.org/0000-0002-5310-2372, NA, NA, NA, NA, https://orcid.org/0000-0003-3703-7651, https://orcid.org/0000-0001-5055-0316, https://orcid.org/0000-0003-4462-1020, https://openalex.org/I170897317, Duke University, https://ror.org/00py81415, US, education, https://openalex.org/I170897317, Duke University, https://ror.org/00py81415, US, education, https://openalex.org/I170897317, Duke University, https://ror.org/00py81415, US, education, https://openalex.org/I170897317, Duke University, https://ror.org/00py81415, US, education, https://openalex.org/I170897317, Duke University, https://ror.org/00py81415, US, education, https://openalex.org/I124055696, https://openalex.org/I170897317, University of Copenhagen, Duke University, https://ror.org/035b05819, https://ror.org/00py81415, DK, US, education, education, https://openalex.org/I114027177, University of North Carolina at Chapel Hill, https://ror.org/0130frc33, US, education, https://openalex.org/I170897317, Duke University, https://ror.org/00py81415, US, education, first, middle, middle, middle, middle, middle, middle, last, Duke University#TAB#, Duke University#TAB#, Duke University#TAB#, Duke University#TAB#, Duke University#TAB#, University of Copenhagen, University of North Carolina at Chapel Hill#TAB#, Duke University#TAB#
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     https://openalex.org/A1069028179, https://openalex.org/A2941293769, William Carbonaro, Emily Maloney, https://orcid.org/0000-0002-5943-3850, NA, https://openalex.org/I107639228, University of Notre Dame, https://ror.org/00mkhxb43, US, education, https://openalex.org/I170897317, Duke University, https://ror.org/00py81415, US, education, first, last, University of Notre Dame, Notre Dame, IN,USA, Duke University, Durham, NC, USA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 concepts
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                              https://openalex.org/C15744967, https://openalex.org/C77805123, https://openalex.org/C130064352, https://openalex.org/C2780597233, https://openalex.org/C144024400, Psychology, Social psychology, Social relation, Deference, Sociology, 0.575016, 0.560851, 0.367378, 0.329392, 0.328239, 0, 1, 2, 2, 0, https://www.wikidata.org/wiki/Q9418, https://www.wikidata.org/wiki/Q161272, https://www.wikidata.org/wiki/Q853725, https://www.wikidata.org/wiki/Q396568, https://www.wikidata.org/wiki/Q21201
#> 2 https://openalex.org/C518677369, https://openalex.org/C110875604, https://openalex.org/C94625758, https://openalex.org/C17744445, https://openalex.org/C39549134, https://openalex.org/C108170787, https://openalex.org/C90508534, https://openalex.org/C108827166, https://openalex.org/C15744967, Social media, The Internet, Politics, Political science, Public relations, Agency (philosophy), Political communication, Internet privacy, Psychology, 0.524221, 0.480312, 0.462947, 0.439907, 0.391009, 0.386419, 0.361514, 0.356026, 0.345458, 2, 2, 2, 0, 1, 2, 3, 1, 0, https://www.wikidata.org/wiki/Q202833, https://www.wikidata.org/wiki/Q75, https://www.wikidata.org/wiki/Q7163, https://www.wikidata.org/wiki/Q36442, https://www.wikidata.org/wiki/Q133080, https://www.wikidata.org/wiki/Q3951828, https://www.wikidata.org/wiki/Q1660049, https://www.wikidata.org/wiki/Q175975, https://www.wikidata.org/wiki/Q9418
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     https://openalex.org/C15744967, https://openalex.org/C145420912, https://openalex.org/C2781206393, Psychology, Mathematics education, Academic achievement, 0.64956, 0.4708, 0.347366, 0, 1, 2, https://www.wikidata.org/wiki/Q9418, https://www.wikidata.org/wiki/Q853077, https://www.wikidata.org/wiki/Q2748419
#>                                                                                                                                                                                                                                                                                                                                                                                              alternate_host_venues
#> 1                                                                                                                                                                                                                                                                                                                                                                                                             NULL
#> 2 https://openalex.org/V125754415, NA, NA, Proceedings of the National Academy of Sciences of the United States of America, Europe PMC, PMC, TRUE, TRUE, TRUE, cc-by-nc-nd, cc-by-nc-nd, NA, https://www.pnas.org/content/pnas/117/1/243.full.pdf, http://europepmc.org/articles/pmc6955293?pdf=render, https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6955293, publishedVersion, publishedVersion, publishedVersion
#> 3                                                                                                                                                                                                                                                                                                          NA, NA, TRUE, cc-by-nc, https://journals.sagepub.com/doi/pdf/10.1177/2378023119845496, publishedVersion
#>   open_access.is_oa open_access.oa_status
#> 1              TRUE                  <NA>
#> 2              TRUE                hybrid
#> 3              TRUE                  gold
#>                                              open_access.oa_url cited_by_count
#> 1                                                          <NA>              0
#> 2          https://www.pnas.org/content/pnas/117/1/243.full.pdf             44
#> 3 https://journals.sagepub.com/doi/pdf/10.1177/2378023119845496              3
#>                                      ids.doi    ids.mag
#> 1 https://doi.org/10.25384/sage.c.5054390.v1 3041426895
#> 2    https://doi.org/10.1073/pnas.1906420116 2991581987
#> 3   https://doi.org/10.1177/2378023119845496 2941505258
#>                       ids.openalex                                 ids.pmid
#> 1 https://openalex.org/W3041426895                                     <NA>
#> 2 https://openalex.org/W2991581987 https://pubmed.ncbi.nlm.nih.gov/31767743
#> 3 https://openalex.org/W2941505258                                     <NA>
#>   publication_year                                        cited_by_api_url
#> 1             2020 https://api.openalex.org/works?filter=cites:W3041426895
#> 2             2020 https://api.openalex.org/works?filter=cites:W2991581987
#> 3             2019 https://api.openalex.org/works?filter=cites:W2941505258
#>                                          doi            type is_paratext
#> 1 https://doi.org/10.25384/sage.c.5054390.v1 journal-article          NA
#> 2    https://doi.org/10.1073/pnas.1906420116 journal-article       FALSE
#> 3   https://doi.org/10.1177/2378023119845496 journal-article       FALSE
#>   is_retracted
#> 1        FALSE
#> 2        FALSE
#> 3        FALSE
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         referenced_works
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   https://openalex.org/W2124242630, https://openalex.org/W2118690925, https://openalex.org/W1976245924, https://openalex.org/W2026926590, https://openalex.org/W2016381774, https://openalex.org/W2464393218, https://openalex.org/W2801576063, https://openalex.org/W2560826082, https://openalex.org/W1980770347, https://openalex.org/W2057392782, https://openalex.org/W2533845211
#> 2                                                                                                                                                                                                                                                                                                                                                                                       https://openalex.org/W2608762407, https://openalex.org/W2937129497, https://openalex.org/W2157979842, https://openalex.org/W1686065478, https://openalex.org/W2888571776, https://openalex.org/W2725249225, https://openalex.org/W2739705275, https://openalex.org/W2079955265, https://openalex.org/W2903903405, https://openalex.org/W2910027323, https://openalex.org/W2912709031, https://openalex.org/W2900936184, https://openalex.org/W2970610561, https://openalex.org/W2904943091, https://openalex.org/W3004404638, https://openalex.org/W2137998590, https://openalex.org/W2161834943, https://openalex.org/W2159397589, https://openalex.org/W2955772761, https://openalex.org/W2947377394, https://openalex.org/W3124643468, https://openalex.org/W3189078335, https://openalex.org/W2889127731, https://openalex.org/W2950230202, https://openalex.org/W2797202416
#> 3 https://openalex.org/W2162491292, https://openalex.org/W2528564138, https://openalex.org/W2086044761, https://openalex.org/W2159434531, https://openalex.org/W2076241933, https://openalex.org/W2072863730, https://openalex.org/W2343715481, https://openalex.org/W2886140641, https://openalex.org/W2019154293, https://openalex.org/W2105596675, https://openalex.org/W2097446348, https://openalex.org/W2114023457, https://openalex.org/W2109046676, https://openalex.org/W2145065148, https://openalex.org/W2113440921, https://openalex.org/W2122350960, https://openalex.org/W2034955177, https://openalex.org/W2170467533, https://openalex.org/W2155230309, https://openalex.org/W2171009698, https://openalex.org/W1553175946, https://openalex.org/W2165185736, https://openalex.org/W2528465614, https://openalex.org/W2151439801, https://openalex.org/W2111954302, https://openalex.org/W1748508028, https://openalex.org/W2000187903, https://openalex.org/W1702295655, https://openalex.org/W2040189309, https://openalex.org/W1963850909, https://openalex.org/W1989327017, https://openalex.org/W2011552753, https://openalex.org/W2135731612, https://openalex.org/W2160440148, https://openalex.org/W3122197175, https://openalex.org/W2019995705
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            related_works
#> 1                                                                                                                                                                                                              https://openalex.org/W2025957539, https://openalex.org/W2153769295, https://openalex.org/W2952430541, https://openalex.org/W2164131384, https://openalex.org/W1828392703, https://openalex.org/W161131878, https://openalex.org/W2111402499, https://openalex.org/W3145542414, https://openalex.org/W2607840051, https://openalex.org/W2488297747, https://openalex.org/W2124489942, https://openalex.org/W3202979141, https://openalex.org/W2800626716, https://openalex.org/W2418216683
#> 2 https://openalex.org/W2913649461, https://openalex.org/W2790166049, https://openalex.org/W3125182500, https://openalex.org/W2910027323, https://openalex.org/W3015794574, https://openalex.org/W2232384272, https://openalex.org/W2804365752, https://openalex.org/W3004825831, https://openalex.org/W2898745023, https://openalex.org/W2903903405, https://openalex.org/W3103318202, https://openalex.org/W2791544114, https://openalex.org/W2888571776, https://openalex.org/W2724523750, https://openalex.org/W2900936184, https://openalex.org/W2529479612, https://openalex.org/W2898982118, https://openalex.org/W2885316213, https://openalex.org/W2550819555, https://openalex.org/W2582561810
#> 3 https://openalex.org/W2222103148, https://openalex.org/W2004885277, https://openalex.org/W2088093348, https://openalex.org/W2122350960, https://openalex.org/W1999200559, https://openalex.org/W1959934967, https://openalex.org/W2151439801, https://openalex.org/W2868066336, https://openalex.org/W2155230309, https://openalex.org/W2271265805, https://openalex.org/W2160770295, https://openalex.org/W2790769440, https://openalex.org/W2112724940, https://openalex.org/W2961143456, https://openalex.org/W2951067112, https://openalex.org/W2558097971, https://openalex.org/W2528564138, https://openalex.org/W2427406780, https://openalex.org/W2100198006, https://openalex.org/W2085278444
#>   biblio.issue biblio.first_page biblio.volume biblio.last_page
#> 1            4               463            83              475
#> 2            1               243           117              250
#> 3         <NA>   237802311984549             5             <NA>
#>   abstract_inverted_index.Affect abstract_inverted_index.Control
#> 1                              0                               1
#> 2                           NULL                            NULL
#> 3                           NULL                            NULL
#>   abstract_inverted_index.Theory abstract_inverted_index.(ACT)
#> 1                              2                             3
#> 2                           NULL                          NULL
#> 3                           NULL                          NULL
#>   abstract_inverted_index.can abstract_inverted_index.predict
#> 1                       4, 18                               5
#> 2                        NULL                            NULL
#> 3                        NULL                            NULL
#>   abstract_inverted_index.the abstract_inverted_index.average
#> 1                           6                               7
#> 2     19, 29, 43, 59, 79, 188                            NULL
#> 3                        NULL                            NULL
#>   abstract_inverted_index.deference abstract_inverted_index.that
#> 1                                 8                            9
#> 2                              NULL          4, 57, 96, 118, 150
#> 3                              NULL                           20
#>   abstract_inverted_index.occupational abstract_inverted_index.identities
#> 1                               10, 20                                 11
#> 2                                 NULL                               NULL
#> 3                                 NULL                               NULL
#>   abstract_inverted_index.receive abstract_inverted_index.from
#> 1                              12                           13
#> 2                            NULL                       71, 85
#> 3                            NULL                         NULL
#>   abstract_inverted_index.others. abstract_inverted_index.These
#> 1                              14                            15
#> 2                            NULL                          NULL
#> 3                            NULL                          NULL
#>   abstract_inverted_index.“deference abstract_inverted_index.scores”
#> 1                                 16                              17
#> 2                               NULL                            NULL
#> 3                               NULL                            NULL
#>   abstract_inverted_index.capture abstract_inverted_index.status
#> 1                              19                             21
#> 2                            NULL                           NULL
#> 3                            NULL                           NULL
#>   abstract_inverted_index.better abstract_inverted_index.than
#> 1                             22                           23
#> 2                           NULL                         NULL
#> 3                           NULL                         NULL
#>   abstract_inverted_index.previous abstract_inverted_index.operati...
#> 1                               24                                 25
#> 2                             NULL                               NULL
#> 3                             NULL                               NULL
#>   abstract_inverted_index.There abstract_inverted_index.is
#> 1                          NULL                       NULL
#> 2                             0                      1, 35
#> 3                          NULL                          5
#>   abstract_inverted_index.widespread abstract_inverted_index.concern
#> 1                               NULL                            NULL
#> 2                                  2                               3
#> 3                               NULL                            NULL
#>   abstract_inverted_index.Russia  abstract_inverted_index.and
#> 1                           NULL                         NULL
#> 2                              5 6, 46, 61, 67, 109, 140, 206
#> 3                           NULL                       10, 15
#>   abstract_inverted_index.other abstract_inverted_index.countries
#> 1                          NULL                              NULL
#> 2                             7                                 8
#> 3                          NULL                              NULL
#>   abstract_inverted_index.have abstract_inverted_index.launched
#> 1                         NULL                             NULL
#> 2                       9, 154                               10
#> 3                         NULL                             NULL
#>   abstract_inverted_index.social-media abstract_inverted_index.campaigns
#> 1                                 NULL                              NULL
#> 2                                   11                                12
#> 3                                 NULL                              NULL
#>   abstract_inverted_index.designed abstract_inverted_index.to
#> 1                             NULL                       NULL
#> 2                               13               14, 156, 182
#> 3                             NULL                          8
#>   abstract_inverted_index.increase abstract_inverted_index.political
#> 1                             NULL                              NULL
#> 2                               15                  16, 44, 107, 204
#> 3                             NULL                              NULL
#>   abstract_inverted_index.divisions abstract_inverted_index.in
#> 1                              NULL                       NULL
#> 2                                17                    18, 138
#> 3                              NULL                          1
#>   abstract_inverted_index.United abstract_inverted_index.States.
#> 1                           NULL                            NULL
#> 2                             20                              21
#> 3                           NULL                            NULL
#>   abstract_inverted_index.Though abstract_inverted_index.a
#> 1                           NULL                      NULL
#> 2                             22                   23, 112
#> 3                           NULL                      NULL
#>   abstract_inverted_index.growing abstract_inverted_index.number
#> 1                            NULL                           NULL
#> 2                              24                             25
#> 3                            NULL                           NULL
#>      abstract_inverted_index.of abstract_inverted_index.studies
#> 1                          NULL                            NULL
#> 2 26, 31, 48, 64, 106, 143, 177                              27
#> 3                            22                            NULL
#>   abstract_inverted_index.analyze abstract_inverted_index.strategy
#> 1                            NULL                             NULL
#> 2                              28                               30
#> 3                            NULL                             NULL
#>   abstract_inverted_index.such abstract_inverted_index.campaigns,
#> 1                         NULL                               NULL
#> 2                           32                            33, 203
#> 3                         NULL                               NULL
#>   abstract_inverted_index.it abstract_inverted_index.not
#> 1                       NULL                        NULL
#> 2                         34                          36
#> 3                       NULL                        NULL
#>   abstract_inverted_index.yet abstract_inverted_index.known
#> 1                        NULL                          NULL
#> 2                          37                            38
#> 3                        NULL                          NULL
#>   abstract_inverted_index.how abstract_inverted_index.these
#> 1                        NULL                          NULL
#> 2                          39                       40, 147
#> 3                        NULL                          NULL
#>   abstract_inverted_index.efforts abstract_inverted_index.shaped
#> 1                            NULL                           NULL
#> 2                              41                             42
#> 3                            NULL                           NULL
#>   abstract_inverted_index.attitudes abstract_inverted_index.behaviors
#> 1                              NULL                              NULL
#> 2                       45, 60, 108                       47, 63, 110
#> 3                              NULL                              NULL
#>   abstract_inverted_index.Americans. abstract_inverted_index.We
#> 1                               NULL                       NULL
#> 2                                 49               50, 115, 170
#> 3                               NULL                       NULL
#>   abstract_inverted_index.study abstract_inverted_index.this
#> 1                          NULL                         NULL
#> 2                            51                           52
#> 3                          NULL                         NULL
#>   abstract_inverted_index.question abstract_inverted_index.using
#> 1                             NULL                          NULL
#> 2                               53                            54
#> 3                             NULL                          NULL
#>   abstract_inverted_index.longitudinal abstract_inverted_index.data
#> 1                                 NULL                         NULL
#> 2                                   55                       56, 77
#> 3                                 NULL                         NULL
#>   abstract_inverted_index.describe abstract_inverted_index.online
#> 1                             NULL                           NULL
#> 2                               58                             62
#> 3                             NULL                           NULL
#>   abstract_inverted_index.1,239 abstract_inverted_index.Republican
#> 1                          NULL                               NULL
#> 2                            65                                 66
#> 3                          NULL                               NULL
#>   abstract_inverted_index.Democratic abstract_inverted_index.Twitter
#> 1                               NULL                            NULL
#> 2                                 68                    69, 134, 144
#> 3                               NULL                            NULL
#>   abstract_inverted_index.users abstract_inverted_index.late
#> 1                          NULL                         NULL
#> 2                            70                           72
#> 3                          NULL                         NULL
#>   abstract_inverted_index.2017 abstract_inverted_index.merged
#> 1                         NULL                           NULL
#> 2                           73                             74
#> 3                         NULL                           NULL
#>   abstract_inverted_index.with abstract_inverted_index.nonpublic
#> 1                         NULL                              NULL
#> 2        75, 98, 120, 128, 163                                76
#> 3                         NULL                              NULL
#>   abstract_inverted_index.about abstract_inverted_index.Russian
#> 1                          NULL                            NULL
#> 2                            78                         80, 151
#> 3                          NULL                            NULL
#>   abstract_inverted_index.Internet abstract_inverted_index.Research
#> 1                             NULL                             NULL
#> 2                               81                               82
#> 3                             NULL                             NULL
#>   abstract_inverted_index.Agency abstract_inverted_index.(IRA)
#> 1                           NULL                          NULL
#> 2                             83                            84
#> 3                           NULL                          NULL
#>   abstract_inverted_index.Twitter. abstract_inverted_index.Using
#> 1                             NULL                          NULL
#> 2                               86                            87
#> 3                             NULL                          NULL
#>   abstract_inverted_index.Bayesian abstract_inverted_index.regression
#> 1                             NULL                               NULL
#> 2                               88                                 89
#> 3                             NULL                               NULL
#>   abstract_inverted_index.tree abstract_inverted_index.models,
#> 1                         NULL                            NULL
#> 2                           90                              91
#> 3                         NULL                            NULL
#>   abstract_inverted_index.we abstract_inverted_index.find
#> 1                       NULL                         NULL
#> 2                         92                      93, 117
#> 3                       NULL                         NULL
#>   abstract_inverted_index.no abstract_inverted_index.evidence
#> 1                       NULL                             NULL
#> 2                         94                               95
#> 3                       NULL                             NULL
#>   abstract_inverted_index.interaction abstract_inverted_index.IRA
#> 1                                NULL                        NULL
#> 2                             97, 119                99, 121, 185
#> 3                                NULL                        NULL
#>   abstract_inverted_index.accounts abstract_inverted_index.substantially
#> 1                             NULL                                  NULL
#> 2                    100, 122, 186                                   101
#> 3                             NULL                                  NULL
#>   abstract_inverted_index.impacted abstract_inverted_index.6
#> 1                             NULL                      NULL
#> 2                              102                       103
#> 3                             NULL                      NULL
#>   abstract_inverted_index.distinctive abstract_inverted_index.measures
#> 1                                NULL                             NULL
#> 2                                 104                              105
#> 3                                NULL                             NULL
#>   abstract_inverted_index.over abstract_inverted_index.1-mo
#> 1                         NULL                         NULL
#> 2                          111                          113
#> 3                         NULL                         NULL
#>   abstract_inverted_index.period. abstract_inverted_index.also
#> 1                            NULL                         NULL
#> 2                             114                          116
#> 3                            NULL                         NULL
#>   abstract_inverted_index.were abstract_inverted_index.most
#> 1                         NULL                         NULL
#> 2                     123, 166                          124
#> 3                         NULL                         NULL
#>   abstract_inverted_index.common abstract_inverted_index.among
#> 1                           NULL                          NULL
#> 2                            125                           126
#> 3                           NULL                          NULL
#>   abstract_inverted_index.respondents abstract_inverted_index.strong
#> 1                                NULL                           NULL
#> 2                                 127                            129
#> 3                                NULL                           NULL
#>   abstract_inverted_index.ideological abstract_inverted_index.homophily
#> 1                                NULL                              NULL
#> 2                                 130                               131
#> 3                                NULL                              NULL
#>   abstract_inverted_index.within abstract_inverted_index.their
#> 1                           NULL                          NULL
#> 2                            132                           133
#> 3                           NULL                          NULL
#>   abstract_inverted_index.network, abstract_inverted_index.high
#> 1                             NULL                         NULL
#> 2                              135                     136, 141
#> 3                             NULL                         NULL
#>   abstract_inverted_index.interest abstract_inverted_index.politics,
#> 1                             NULL                              NULL
#> 2                              137                               139
#> 3                             NULL                              NULL
#>   abstract_inverted_index.frequency abstract_inverted_index.usage.
#> 1                              NULL                           NULL
#> 2                               142                            145
#> 3                              NULL                           NULL
#>   abstract_inverted_index.Together, abstract_inverted_index.findings
#> 1                              NULL                             NULL
#> 2                               146                              148
#> 3                              NULL                             NULL
#>   abstract_inverted_index.suggest abstract_inverted_index.trolls
#> 1                            NULL                           NULL
#> 2                             149                            152
#> 3                            NULL                           NULL
#>   abstract_inverted_index.might abstract_inverted_index.failed
#> 1                          NULL                           NULL
#> 2                           153                            155
#> 3                          NULL                           NULL
#>   abstract_inverted_index.sow abstract_inverted_index.discord
#> 1                        NULL                            NULL
#> 2                         157                             158
#> 3                        NULL                            NULL
#>   abstract_inverted_index.because abstract_inverted_index.they
#> 1                            NULL                         NULL
#> 2                             159                          160
#> 3                              21                         NULL
#>   abstract_inverted_index.mostly abstract_inverted_index.interacted
#> 1                           NULL                               NULL
#> 2                            161                                162
#> 3                           NULL                               NULL
#>   abstract_inverted_index.those abstract_inverted_index.who
#> 1                          NULL                        NULL
#> 2                           164                         165
#> 3                          NULL                        NULL
#>   abstract_inverted_index.already abstract_inverted_index.highly
#> 1                            NULL                           NULL
#> 2                             167                            168
#> 3                            NULL                           NULL
#>   abstract_inverted_index.polarized. abstract_inverted_index.conclude
#> 1                               NULL                             NULL
#> 2                                169                              171
#> 3                               NULL                             NULL
#>   abstract_inverted_index.by abstract_inverted_index.discussing
#> 1                       NULL                               NULL
#> 2                        172                                173
#> 3                       NULL                               NULL
#>   abstract_inverted_index.several abstract_inverted_index.important
#> 1                            NULL                              NULL
#> 2                             174                               175
#> 3                            NULL                              NULL
#>   abstract_inverted_index.limitations abstract_inverted_index.our
#> 1                                NULL                        NULL
#> 2                                 176                    178, 180
#> 3                                NULL                        NULL
#>   abstract_inverted_index.study—especially abstract_inverted_index.inability
#> 1                                     NULL                              NULL
#> 2                                      179                               181
#> 3                                     NULL                              NULL
#>   abstract_inverted_index.determine abstract_inverted_index.whether
#> 1                              NULL                            NULL
#> 2                               183                             184
#> 3                              NULL                            NULL
#>   abstract_inverted_index.influenced abstract_inverted_index.2016
#> 1                               NULL                         NULL
#> 2                                187                          189
#> 3                               NULL                         NULL
#>   abstract_inverted_index.presidential abstract_inverted_index.election—as
#> 1                                 NULL                                NULL
#> 2                                  190                                 191
#> 3                                 NULL                                NULL
#>   abstract_inverted_index.well abstract_inverted_index.as
#> 1                         NULL                       NULL
#> 2                          192                        193
#> 3                         NULL                       NULL
#>   abstract_inverted_index.its abstract_inverted_index.implications
#> 1                        NULL                                 NULL
#> 2                         194                                  195
#> 3                        NULL                                 NULL
#>   abstract_inverted_index.for abstract_inverted_index.future
#> 1                        NULL                           NULL
#> 2                         196                            197
#> 3                          13                           NULL
#>   abstract_inverted_index.research abstract_inverted_index.on
#> 1                             NULL                       NULL
#> 2                              198                        199
#> 3                             NULL                       NULL
#>   abstract_inverted_index.social abstract_inverted_index.media
#> 1                           NULL                          NULL
#> 2                       200, 208                           201
#> 3                           NULL                          NULL
#>   abstract_inverted_index.influence abstract_inverted_index.polarization,
#> 1                              NULL                                  NULL
#> 2                               202                                   205
#> 3                              NULL                                  NULL
#>   abstract_inverted_index.computational abstract_inverted_index.science.
#> 1                                  NULL                             NULL
#> 2                                   207                              209
#> 3                                  NULL                             NULL
#>   abstract_inverted_index.Participation abstract_inverted_index.extracurricular
#> 1                                  NULL                                    NULL
#> 2                                  NULL                                    NULL
#> 3                                     0                                       2
#>   abstract_inverted_index.activities abstract_inverted_index.(ECAs)
#> 1                               NULL                           NULL
#> 2                               NULL                           NULL
#> 3                                  3                              4
#>   abstract_inverted_index.positively abstract_inverted_index.related
#> 1                               NULL                            NULL
#> 2                               NULL                            NULL
#> 3                                  6                               7
#>   abstract_inverted_index.cognitive abstract_inverted_index.socioemotional
#> 1                              NULL                                   NULL
#> 2                              NULL                                   NULL
#> 3                                 9                                     11
#>   abstract_inverted_index.outcomes abstract_inverted_index.children
#> 1                             NULL                             NULL
#> 2                             NULL                             NULL
#> 3                               12                               14
#>   abstract_inverted_index.adolescents. abstract_inverted_index.The
#> 1                                 NULL                        NULL
#> 2                                 NULL                        NULL
#> 3                                   16                          17
#>   abstract_inverted_index.authors abstract_inverted_index.argue
#> 1                            NULL                          NULL
#> 2                            NULL                          NULL
#> 3                              18                            19
#>   abstract_inverted_index.methodological abstract_inverted_index.limi...
#> 1                                   NULL                            NULL
#> 2                                   NULL                            NULL
#> 3                                     23                              24