openalex_api

Baseline function for calling the API and parsing results

Usage

openalex_api(path)

Arguments

path

the url path to use in the GET request to the openalex API

Value

list of parsed JSON returned from the API

Examples

openalex_api(path = "https://api.openalex.org/works/W2741809807")
#> $id
#> [1] "https://openalex.org/W2741809807"
#> 
#> $doi
#> [1] "https://doi.org/10.7717/peerj.4375"
#> 
#> $display_name
#> [1] "The state of OA: a large-scale analysis of the prevalence and impact of Open Access articles"
#> 
#> $title
#> [1] "The state of OA: a large-scale analysis of the prevalence and impact of Open Access articles"
#> 
#> $publication_year
#> [1] 2018
#> 
#> $publication_date
#> [1] "2018-02-13"
#> 
#> $ids
#> $ids$openalex
#> [1] "https://openalex.org/W2741809807"
#> 
#> $ids$doi
#> [1] "https://doi.org/10.7717/peerj.4375"
#> 
#> $ids$pmid
#> [1] "https://pubmed.ncbi.nlm.nih.gov/29456894"
#> 
#> $ids$mag
#> [1] 2741809807
#> 
#> 
#> $host_venue
#> $host_venue$id
#> [1] "https://openalex.org/V1983995261"
#> 
#> $host_venue$issn_l
#> [1] "2167-8359"
#> 
#> $host_venue$issn
#> [1] "2167-8359"
#> 
#> $host_venue$display_name
#> [1] "PeerJ"
#> 
#> $host_venue$publisher
#> [1] "PeerJ"
#> 
#> $host_venue$type
#> [1] "publisher"
#> 
#> $host_venue$url
#> [1] "https://doi.org/10.7717/peerj.4375"
#> 
#> $host_venue$is_oa
#> [1] TRUE
#> 
#> $host_venue$version
#> [1] "publishedVersion"
#> 
#> $host_venue$license
#> [1] "cc-by"
#> 
#> 
#> $type
#> [1] "journal-article"
#> 
#> $open_access
#> $open_access$is_oa
#> [1] TRUE
#> 
#> $open_access$oa_status
#> [1] "gold"
#> 
#> $open_access$oa_url
#> [1] "https://peerj.com/articles/4375.pdf"
#> 
#> 
#> $authorships
#>   author_position                        author.id author.display_name
#> 1           first https://openalex.org/A1969205032  Heather A. Piwowar
#> 2          middle https://openalex.org/A2208157607         Jason Priem
#> 3          middle  https://openalex.org/A291762745   Vincent Larivière
#> 4          middle https://openalex.org/A2790141563  Juan Pablo Alperin
#> 5          middle https://openalex.org/A2738039797       Lisa Matthias
#> 6          middle https://openalex.org/A2741312048      Bree Norlander
#> 7          middle https://openalex.org/A2741153153       Ashley Farley
#> 8          middle https://openalex.org/A2162292103       Jevin D. West
#> 9            last https://openalex.org/A2027010983   Stefanie Haustein
#>                            author.orcid
#> 1 https://orcid.org/0000-0003-1613-5981
#> 2 https://orcid.org/0000-0001-6187-6610
#> 3 https://orcid.org/0000-0002-2733-0689
#> 4 https://orcid.org/0000-0002-9344-7439
#> 5 https://orcid.org/0000-0002-2612-2132
#> 6 https://orcid.org/0000-0002-0431-4221
#> 7                                  <NA>
#> 8 https://orcid.org/0000-0002-4118-0322
#> 9 https://orcid.org/0000-0003-0157-1430
#>                                                                                                                                                                                                   institutions
#> 1                                                                                                                      https://openalex.org/I4200000001, OurResearch, https://ror.org/02nr0ka47, CA, nonprofit
#> 2                                                                                                                      https://openalex.org/I4200000001, OurResearch, https://ror.org/02nr0ka47, CA, nonprofit
#> 3 https://openalex.org/I70931966, https://openalex.org/I159129438, Université de Montréal, Université du Québec à Montréal, https://ror.org/0161xgx34, https://ror.org/002rjbv21, CA, CA, education, education
#> 4                                     https://openalex.org/I18014758, https://openalex.org/I209863525, Simon Fraser University, Public Knowledge Project, https://ror.org/0213rcc28, NA, CA, NA, education, NA
#> 5                                                                                                            https://openalex.org/I18014758, Simon Fraser University, https://ror.org/0213rcc28, CA, education
#> 6                                                                                                          https://openalex.org/I201448701, University of Washington, https://ror.org/00cvxb145, US, education
#> 7                                                                                                          https://openalex.org/I201448701, University of Washington, https://ror.org/00cvxb145, US, education
#> 8                                                                                                          https://openalex.org/I201448701, University of Washington, https://ror.org/00cvxb145, US, education
#> 9  https://openalex.org/I153718931, https://openalex.org/I159129438, University of Ottawa, Université du Québec à Montréal, https://ror.org/03c4mmv16, https://ror.org/002rjbv21, CA, CA, education, education
#>                                                                                      raw_affiliation_string
#> 1                                                                             Impactstory, Sanford, NC, USA
#> 2                                                                             Impactstory, Sanford, NC, USA
#> 3 École de bibliothéconomie et des sciences de l'information, Université de Montréal, Montréal, QC, Canada.
#> 4             Canadian Institute for Studies in Publishing, Simon Fraser University, Vancouver, BC, Canada.
#> 5                                 Scholarly Communications Lab, Simon Fraser University, Vancouver, Canada.
#> 6                                                Information School, University of Washington, Seattle, USA
#> 7                                                Information School, University of Washington, Seattle, USA
#> 8                                              [Information School, University of Washington, Seattle, USA]
#> 9                                   School of Information Studies, University of Ottawa, Ottawa, ON, Canada
#> 
#> $cited_by_count
#> [1] 382
#> 
#> $biblio
#> $biblio$volume
#> [1] "6"
#> 
#> $biblio$issue
#> NULL
#> 
#> $biblio$first_page
#> NULL
#> 
#> $biblio$last_page
#> NULL
#> 
#> 
#> $is_retracted
#> [1] FALSE
#> 
#> $is_paratext
#> [1] FALSE
#> 
#> $concepts
#>                                  id                               wikidata
#> 1  https://openalex.org/C2778793908 https://www.wikidata.org/wiki/Q5122404
#> 2  https://openalex.org/C2778805511    https://www.wikidata.org/wiki/Q1713
#> 3   https://openalex.org/C178315738  https://www.wikidata.org/wiki/Q603441
#> 4  https://openalex.org/C2777462167 https://www.wikidata.org/wiki/Q7432048
#> 5   https://openalex.org/C525823164  https://www.wikidata.org/wiki/Q472342
#> 6  https://openalex.org/C2778149293  https://www.wikidata.org/wiki/Q309823
#> 7  https://openalex.org/C2780560020   https://www.wikidata.org/wiki/Q79719
#> 8   https://openalex.org/C151719136 https://www.wikidata.org/wiki/Q3972943
#> 9    https://openalex.org/C74909509   https://www.wikidata.org/wiki/Q10387
#> 10  https://openalex.org/C161191863  https://www.wikidata.org/wiki/Q199655
#> 11   https://openalex.org/C71924100   https://www.wikidata.org/wiki/Q11190
#>               display_name level    score
#> 1          Citation impact     3 0.459309
#> 2                 Citation     2 0.447306
#> 3            Bibliometrics     2 0.444802
#> 4  Scholarly communication     3 0.441050
#> 5           Scientometrics     2 0.430134
#> 6             Open science     2 0.424984
#> 7                  License     2 0.405121
#> 8               Publishing     2 0.403990
#> 9              Gerontology     1 0.396626
#> 10         Library science     1 0.387624
#> 11                Medicine     0 0.352338
#> 
#> $mesh
#> list()
#> 
#> $alternate_host_venues
#>                                 id
#> 1 https://openalex.org/V1983995261
#> 2                             <NA>
#> 3                             <NA>
#> 4                             <NA>
#> 5                             <NA>
#> 6                             <NA>
#> 7                             <NA>
#> 8                             <NA>
#>                                                  display_name       type
#> 1                                                       PeerJ    journal
#> 2                                                  Europe PMC repository
#> 3                                                         PMC repository
#> 4                            Simon Fraser University - Summit repository
#> 5 Université de Montréal - Papyrus : Institutional Repository repository
#> 6                            Simon Fraser University - Summit repository
#> 7                                                        <NA> repository
#> 8                                            Semantic Scholar repository
#>                                                                                url
#> 1                                              https://peerj.com/articles/4375.pdf
#> 2                              http://europepmc.org/articles/pmc5815332?pdf=render
#> 3                             https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5815332
#> 4                                                 https://summit.sfu.ca/item/17691
#> 5  https://papyrus.bib.umontreal.ca/xmlui/bitstream/1866/23242/1/peerj-06-4375.pdf
#> 6                  http://summit.sfu.ca/system/files/iritems1/17691/peerj-4375.pdf
#> 7 https://digitalcommons.unl.edu/cgi/viewcontent.cgi?article=1143&context=scholcom
#> 8    http://pdfs.semanticscholar.org/bca1/531b0562c6d72e0c283c1ccb97eb5cb02117.pdf
#>   is_oa          version license
#> 1  TRUE publishedVersion   cc-by
#> 2  TRUE publishedVersion   cc-by
#> 3  TRUE publishedVersion    <NA>
#> 4  TRUE submittedVersion   cc-by
#> 5  TRUE submittedVersion   cc-by
#> 6  TRUE submittedVersion   cc-by
#> 7  TRUE submittedVersion   cc-by
#> 8  TRUE submittedVersion    <NA>
#> 
#> $referenced_works
#>  [1] "https://openalex.org/W2520991028" "https://openalex.org/W3121567788"
#>  [3] "https://openalex.org/W1560783210" "https://openalex.org/W2463568293"
#>  [5] "https://openalex.org/W2953072907" "https://openalex.org/W2511661767"
#>  [7] "https://openalex.org/W2231201268" "https://openalex.org/W2343014812"
#>  [9] "https://openalex.org/W2048185449" "https://openalex.org/W2345375849"
#> [11] "https://openalex.org/W2566143661" "https://openalex.org/W2762597540"
#> [13] "https://openalex.org/W1724212071" "https://openalex.org/W2020807482"
#> [15] "https://openalex.org/W2078310052" "https://openalex.org/W2322381034"
#> [17] "https://openalex.org/W2003844967" "https://openalex.org/W2016860460"
#> [19] "https://openalex.org/W2140880926" "https://openalex.org/W2588027260"
#> [21] "https://openalex.org/W1767272795" "https://openalex.org/W1917633107"
#> [23] "https://openalex.org/W1957687230" "https://openalex.org/W1989318653"
#> [25] "https://openalex.org/W2753353163" "https://openalex.org/W2785823074"
#> [27] "https://openalex.org/W2029057325" "https://openalex.org/W2306268324"
#> [29] "https://openalex.org/W2115339903" "https://openalex.org/W2031754690"
#> [31] "https://openalex.org/W2563251083" "https://openalex.org/W2587705861"
#> [33] "https://openalex.org/W2160597895" "https://openalex.org/W2089123513"
#> 
#> $related_works
#>  [1] "https://openalex.org/W2089604928" "https://openalex.org/W1996515099"
#>  [3] "https://openalex.org/W2587705861" "https://openalex.org/W2302501749"
#>  [5] "https://openalex.org/W2029057325" "https://openalex.org/W2023783748"
#>  [7] "https://openalex.org/W2785823074" "https://openalex.org/W2783957468"
#>  [9] "https://openalex.org/W2016860460" "https://openalex.org/W2463568293"
#> [11] "https://openalex.org/W2048185449" "https://openalex.org/W2089123513"
#> [13] "https://openalex.org/W2104772551" "https://openalex.org/W2520991028"
#> [15] "https://openalex.org/W2962868865" "https://openalex.org/W2140880926"
#> [17] "https://openalex.org/W1970217003" "https://openalex.org/W1989318653"
#> [19] "https://openalex.org/W2322381034" "https://openalex.org/W1655306493"
#> 
#> $abstract_inverted_index
#> $abstract_inverted_index$Despite
#> [1] 0
#> 
#> $abstract_inverted_index$growing
#> [1] 1
#> 
#> $abstract_inverted_index$interest
#> [1] 2
#> 
#> $abstract_inverted_index$`in`
#> [1]   3  57  73 110 122
#> 
#> $abstract_inverted_index$Open
#> [1]   4 201
#> 
#> $abstract_inverted_index$Access
#> [1] 5
#> 
#> $abstract_inverted_index$`(OA)`
#> [1] 6
#> 
#> $abstract_inverted_index$to
#> [1]   7  54 252
#> 
#> $abstract_inverted_index$scholarly
#> [1]   8 105
#> 
#> $abstract_inverted_index$`literature,`
#> [1] 9
#> 
#> $abstract_inverted_index$there
#> [1] 10
#> 
#> $abstract_inverted_index$is
#> [1]  11 107 116 176
#> 
#> $abstract_inverted_index$an
#> [1]  12  34  85 185 199 231
#> 
#> $abstract_inverted_index$unmet
#> [1] 13
#> 
#> $abstract_inverted_index$need
#> [1] 14 31
#> 
#> $abstract_inverted_index$`for`
#> [1]  15  42 174 219
#> 
#> $abstract_inverted_index$`large-scale,`
#> [1] 16
#> 
#> $abstract_inverted_index$`up-to-date,`
#> [1] 17
#> 
#> $abstract_inverted_index$and
#> [1]  18  24  77 112 124 144 221 237 256
#> 
#> $abstract_inverted_index$reproducible
#> [1] 19
#> 
#> $abstract_inverted_index$studies
#> [1] 20
#> 
#> $abstract_inverted_index$assessing
#> [1] 21
#> 
#> $abstract_inverted_index$the
#> [1]  22 104 134 145 170 195 206 213 245
#> 
#> $abstract_inverted_index$prevalence
#> [1] 23
#> 
#> $abstract_inverted_index$characteristics
#> [1] 25
#> 
#> $abstract_inverted_index$of
#> [1]  26  51  75  83 103 137 141 163 209
#> 
#> $abstract_inverted_index$OA.
#> [1]  27 168 239
#> 
#> $abstract_inverted_index$We
#> [1]  28  46  97 203 240
#> 
#> $abstract_inverted_index$address
#> [1] 29
#> 
#> $abstract_inverted_index$this
#> [1]  30 114 142
#> 
#> $abstract_inverted_index$using
#> [1]  32  95 244
#> 
#> $abstract_inverted_index$`oaDOI,`
#> [1] 33
#> 
#> $abstract_inverted_index$open
#> [1] 35
#> 
#> $abstract_inverted_index$online
#> [1] 36
#> 
#> $abstract_inverted_index$service
#> [1] 37
#> 
#> $abstract_inverted_index$that
#> [1]  38  89  99 113 147 155
#> 
#> $abstract_inverted_index$determines
#> [1] 39
#> 
#> $abstract_inverted_index$OA
#>  [1]  40  56  93 108 138 159 175 210 223 254
#> 
#> $abstract_inverted_index$status
#> [1] 41
#> 
#> $abstract_inverted_index$`67`
#> [1] 43
#> 
#> $abstract_inverted_index$million
#> [1] 44
#> 
#> $abstract_inverted_index$articles.
#> [1] 45
#> 
#> $abstract_inverted_index$use
#> [1] 47
#> 
#> $abstract_inverted_index$three
#> [1] 48 58
#> 
#> $abstract_inverted_index$`samples,`
#> [1] 49
#> 
#> $abstract_inverted_index$each
#> [1] 50
#> 
#> $abstract_inverted_index$`100,000`
#> [1] 52
#> 
#> $abstract_inverted_index$`articles,`
#> [1]  53 152 211
#> 
#> $abstract_inverted_index$investigate
#> [1] 55
#> 
#> $abstract_inverted_index$`populations:`
#> [1] 59
#> 
#> $abstract_inverted_index$`(1)`
#> [1] 60
#> 
#> $abstract_inverted_index$all
#> [1] 61
#> 
#> $abstract_inverted_index$journal
#> [1] 62 70
#> 
#> $abstract_inverted_index$articles
#> [1]  63  71  79  94 164 191 224
#> 
#> $abstract_inverted_index$assigned
#> [1] 64
#> 
#> $abstract_inverted_index$a
#> [1]  65 250
#> 
#> $abstract_inverted_index$Crossref
#> [1] 66
#> 
#> $abstract_inverted_index$`DOI,`
#> [1] 67
#> 
#> $abstract_inverted_index$`(2)`
#> [1] 68
#> 
#> $abstract_inverted_index$recent
#> [1]  69 128
#> 
#> $abstract_inverted_index$indexed
#> [1] 72
#> 
#> $abstract_inverted_index$Web
#> [1] 74
#> 
#> $abstract_inverted_index$`Science,`
#> [1] 76
#> 
#> $abstract_inverted_index$`(3)`
#> [1] 78
#> 
#> $abstract_inverted_index$viewed
#> [1] 80
#> 
#> $abstract_inverted_index$by
#> [1]  81 120 235
#> 
#> $abstract_inverted_index$users
#> [1]  82  91 157
#> 
#> $abstract_inverted_index$`Unpaywall,`
#> [1] 84
#> 
#> $abstract_inverted_index$`open-source`
#> [1] 86
#> 
#> $abstract_inverted_index$browser
#> [1] 87
#> 
#> $abstract_inverted_index$extension
#> [1] 88
#> 
#> $abstract_inverted_index$lets
#> [1] 90
#> 
#> $abstract_inverted_index$find
#> [1]  92 154
#> 
#> $abstract_inverted_index$oaDOI.
#> [1] 96
#> 
#> $abstract_inverted_index$estimate
#> [1] 98
#> 
#> $abstract_inverted_index$at
#> [1] 100
#> 
#> $abstract_inverted_index$least
#> [1] 101
#> 
#> $abstract_inverted_index$`28%`
#> [1] 102
#> 
#> $abstract_inverted_index$literature
#> [1] 106
#> 
#> $abstract_inverted_index$`(19M`
#> [1] 109
#> 
#> $abstract_inverted_index$`total)`
#> [1] 111
#> 
#> $abstract_inverted_index$proportion
#> [1] 115
#> 
#> $abstract_inverted_index$`growing,`
#> [1] 117
#> 
#> $abstract_inverted_index$driven
#> [1] 118 233
#> 
#> $abstract_inverted_index$particularly
#> [1] 119
#> 
#> $abstract_inverted_index$growth
#> [1] 121
#> 
#> $abstract_inverted_index$Gold
#> [1] 123
#> 
#> $abstract_inverted_index$Hybrid.
#> [1] 125
#> 
#> $abstract_inverted_index$The
#> [1] 126
#> 
#> $abstract_inverted_index$most
#> [1] 127 171
#> 
#> $abstract_inverted_index$year
#> [1] 129
#> 
#> $abstract_inverted_index$analyzed
#> [1] 130
#> 
#> $abstract_inverted_index$`(2015)`
#> [1] 131
#> 
#> $abstract_inverted_index$also
#> [1] 132 204
#> 
#> $abstract_inverted_index$has
#> [1] 133
#> 
#> $abstract_inverted_index$highest
#> [1] 135
#> 
#> $abstract_inverted_index$percentage
#> [1] 136
#> 
#> $abstract_inverted_index$`(45%).`
#> [1] 139
#> 
#> $abstract_inverted_index$Because
#> [1] 140
#> 
#> $abstract_inverted_index$`growth,`
#> [1] 143
#> 
#> $abstract_inverted_index$fact
#> [1] 146
#> 
#> $abstract_inverted_index$readers
#> [1] 148
#> 
#> $abstract_inverted_index$disproportionately
#> [1] 149
#> 
#> $abstract_inverted_index$access
#> [1] 150
#> 
#> $abstract_inverted_index$newer
#> [1] 151
#> 
#> $abstract_inverted_index$we
#> [1] 153 188
#> 
#> $abstract_inverted_index$Unpaywall
#> [1] 156
#> 
#> $abstract_inverted_index$encounter
#> [1] 158
#> 
#> $abstract_inverted_index$quite
#> [1] 160
#> 
#> $abstract_inverted_index$`frequently:`
#> [1] 161
#> 
#> $abstract_inverted_index$`47%`
#> [1] 162
#> 
#> $abstract_inverted_index$they
#> [1] 165
#> 
#> $abstract_inverted_index$view
#> [1] 166
#> 
#> $abstract_inverted_index$are
#> [1] 167
#> 
#> $abstract_inverted_index$`Notably,`
#> [1] 169
#> 
#> $abstract_inverted_index$common
#> [1] 172
#> 
#> $abstract_inverted_index$mechanism
#> [1] 173
#> 
#> $abstract_inverted_index$not
#> [1] 177
#> 
#> $abstract_inverted_index$`Gold,`
#> [1] 178
#> 
#> $abstract_inverted_index$`Green,`
#> [1] 179
#> 
#> $abstract_inverted_index$or
#> [1] 180
#> 
#> $abstract_inverted_index$Hybrid
#> [1] 181 238
#> 
#> $abstract_inverted_index$`OA,`
#> [1] 182
#> 
#> $abstract_inverted_index$but
#> [1] 183
#> 
#> $abstract_inverted_index$rather
#> [1] 184
#> 
#> $abstract_inverted_index$`under-discussed`
#> [1] 186
#> 
#> $abstract_inverted_index$category
#> [1] 187
#> 
#> $abstract_inverted_index$dub
#> [1] 189
#> 
#> $abstract_inverted_index$`Bronze:`
#> [1] 190
#> 
#> $abstract_inverted_index$made
#> [1] 192
#> 
#> $abstract_inverted_index$`free-to-read`
#> [1] 193
#> 
#> $abstract_inverted_index$on
#> [1] 194
#> 
#> $abstract_inverted_index$publisher
#> [1] 196
#> 
#> $abstract_inverted_index$`website,`
#> [1] 197
#> 
#> $abstract_inverted_index$without
#> [1] 198
#> 
#> $abstract_inverted_index$explicit
#> [1] 200
#> 
#> $abstract_inverted_index$license.
#> [1] 202
#> 
#> $abstract_inverted_index$examine
#> [1] 205
#> 
#> $abstract_inverted_index$citation
#> [1] 207 216
#> 
#> $abstract_inverted_index$impact
#> [1] 208
#> 
#> $abstract_inverted_index$corroborating
#> [1] 212
#> 
#> $abstract_inverted_index$`so-called`
#> [1] 214
#> 
#> $abstract_inverted_index$`open-access`
#> [1] 215
#> 
#> $abstract_inverted_index$`advantage:`
#> [1] 217
#> 
#> $abstract_inverted_index$accounting
#> [1] 218
#> 
#> $abstract_inverted_index$age
#> [1] 220
#> 
#> $abstract_inverted_index$`discipline,`
#> [1] 222
#> 
#> $abstract_inverted_index$receive
#> [1] 225
#> 
#> $abstract_inverted_index$`18%`
#> [1] 226
#> 
#> $abstract_inverted_index$more
#> [1] 227
#> 
#> $abstract_inverted_index$citations
#> [1] 228
#> 
#> $abstract_inverted_index$than
#> [1] 229
#> 
#> $abstract_inverted_index$`average,`
#> [1] 230
#> 
#> $abstract_inverted_index$effect
#> [1] 232
#> 
#> $abstract_inverted_index$primarily
#> [1] 234
#> 
#> $abstract_inverted_index$Green
#> [1] 236
#> 
#> $abstract_inverted_index$encourage
#> [1] 241
#> 
#> $abstract_inverted_index$further
#> [1] 242
#> 
#> $abstract_inverted_index$research
#> [1] 243
#> 
#> $abstract_inverted_index$free
#> [1] 246
#> 
#> $abstract_inverted_index$oaDOI
#> [1] 247
#> 
#> $abstract_inverted_index$`service,`
#> [1] 248
#> 
#> $abstract_inverted_index$as
#> [1] 249
#> 
#> $abstract_inverted_index$way
#> [1] 251
#> 
#> $abstract_inverted_index$inform
#> [1] 253
#> 
#> $abstract_inverted_index$policy
#> [1] 255
#> 
#> $abstract_inverted_index$practice.
#> [1] 257
#> 
#> 
#> $counts_by_year
#>   year cited_by_count
#> 1 2022              4
#> 2 2021            103
#> 3 2020            126
#> 4 2019             97
#> 5 2018             46
#> 6 2017              6
#> 
#> $cited_by_api_url
#> [1] "https://api.openalex.org/works?filter=cites:W2741809807"
#> 
#> $updated_date
#> [1] "2022-01-02T00:22:35.180390"
#>