Simplify the OpenAlex authors result

This function is mostly for the package's internal use, but we export it so you can try it out. However, we expect that you'll likely write your own function to simplify the result however you want.

Usage

show_authors(x, simp_func = utils::head)

Arguments

x

Dataframe/tibble. Result of the OpenAlex query for authors already converted to dataframe/tibble.

simp_func

R function to simplify the result. Default to `head`. If you want the entire table, set `simp_fun = identity`

Value

Simplified tibble to display. The first column, `id` is the short-form OpenAlex ID of the authors.

Examples

show_authors(oa_fetch(
  identifier = c("A923435168", "A2208157607"),
  verbose = TRUE
))
#> Requesting url: https://api.openalex.org/authors?filter=openalex_id%3AA923435168%7CA2208157607
#> Getting 1 page of results with a total of 2 records...
#> # A tibble: 2 × 7
#>   id          display_name orcid               works_c…¹ cited…² affil…³ top_c…⁴
#>   <chr>       <chr>        <chr>                   <int>   <int> <chr>   <chr>  
#> 1 A923435168  Massimo Aria 0000-0002-8517-9411       161    3839 Univer… Statis…
#> 2 A2208157607 Jason Priem  0000-0001-6187-6610        51    1677 HortRe… World …
#> # … with abbreviated variable names ¹​works_count, ²​cited_by_count,
#> #   ³​affiliation_display_name, ⁴​top_concepts