Get Authors Papers

Get Authors Papers

Usage

get_authors_papers(id_type = c("orcid", "openalex", "scopus", "mag"), id)

Arguments

id_type

the type of id you are using to locate the author: must be either orcid or openalex

id

the id for the author that matches the type selected initially

Value

dataframe with all papers from that author, includes variables: id: openalex id of the paper display_name: title of the paper publication date: date of publication in YYYY-MM-DD form relevance score: ?Not sure what this is, all NA for me atm authorships: nested dataframe of all authors in the paper (includes author.id, author_position, author.display_name, and author.orcid) concepts: nested dataframe of the concepts associated with the paper (includes openalex's id of the concept, display.name of the concept, score of the concept, level of the concept in the concept hierarchy, and wikidata of the concept) alternate_host_venues: ?Not sure what this is either, will need to check on openalex's page cited_by_count: number of other papers that have cited the paper publication_year: int cited_by_api_url: openalex api url for the papers that have cited this paper doi: doi of the paper type: type of paper (e.g. proceedings-article, journal-article, book-chapter, posted-content) is_paratext: ?Not sure what this is either, will need to check on openalex's page is_retracted: TRUE/FALSE whether paper has been retracted referenced_works: list of all works referenced in paper (with openalex urls :) ) related_works: similar works? not sure how these are found

Examples

get_authors_papers(id_type = "openalex", id = "A2941293769")
#> # A tibble: 3 × 29
#>   id           doi   mag   openalex pmid  title publication_date relevance_score
#>   <chr>        <chr> <chr> <chr>    <chr> <chr> <chr>            <lgl>          
#> 1 https://ope… http… 3041… https:/… NA    The … 2020-07-08       NA             
#> 2 https://ope… http… 2991… https:/… http… Asse… 2020-01-07       NA             
#> 3 https://ope… http… 2941… https:/… NA    Extr… 2019-04-23       NA             
#> # … with 21 more variables: host_venue_id <chr>, host_venue_name <chr>,
#> #   host_venue_publisher <chr>, host_venue_issn <list>, host_venue_iss_l <chr>,
#> #   host_venue_type <chr>, host_venue_url <chr>, host_venue_oa <lgl>,
#> #   host_venue_version <chr>, host_venue_license <chr>, authorships <list>,
#> #   concepts <list>, alternate_host_venues <list>, cited_by_count <int>,
#> #   publication_year <int>, cited_by_url <chr>, type <chr>, paratext <lgl>,
#> #   retracted <lgl>, referenced_works <list>, related_works <list>