Simplify the OpenAlex authors result

This function is mostly for the package's internal use, but we export it so you can try it out. However, we expect that you'll likely write your own function to simplify the result however you want.

Usage

show_authors(x, simp_func = utils::head)

Arguments

x

Dataframe/tibble. Result of the OpenAlex query for authors already converted to dataframe/tibble.

simp_func

R function to simplify the result. Default to `head`. If you want the entire table, set `simp_fun = identity`

Value

Simplified tibble to display. The first column, `id` is the short-form OpenAlex ID of the authors.

Examples

show_authors(oa_fetch(
  identifier = c("A5023888391", "A5014077037"),
  verbose = TRUE
))
#> Requesting url: https://api.openalex.org/authors?filter=openalex%3AA5023888391%7CA5014077037
#> Getting 1 page of results with a total of 2 records...
#> # A tibble: 2 × 7
#>   id        display_name orcid works_count cited_by_count affiliation_display_…¹
#>   <chr>     <chr>        <chr>       <int>          <int> <chr>                 
#> 1 A5014077… Stefanie Ha… 0000…         215           7200 University of Ottawa  
#> 2 A5023888… Jason Priem  0000…          62           3693 OurResearch           
#> # ℹ abbreviated name: ¹​affiliation_display_name
#> # ℹ 1 more variable: top_concepts <chr>